// src/shared/open-api/models/meResponseDto.ts
 
/**
 * Generated by orval v1.0 🍺
 * Do not edit manually.
 * Micro App API
 */

export interface MeResponseDto {
  id: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  lastActiveWorkspaceId?: string | null;
  /**
   * The primary on-chain wallet address used to identify the profile as an AuthIdentity
   * @nullable
   */
  blockchainAccount?: string | null;
}